+++ packages/component/src/page/trading/sections/chartView/tradeData.tsx
@@ -1,6 +1,7 @@
-import { FC } from "react";
+import { FC, useContext } from "react";
 import { useMarkPrice, useTickerStream } from "@orderly.network/hooks";
 import { Numeral, Text } from "@/text";
+import { SymbolContext } from "@/provider";
 
 interface Props {
   symbol: string;
@@ -8,38 +9,48 @@ interface Props {
 
 export const TradeData: FC<Props> = (props) => {
   const { symbol } = props;
-
-  // const { data: markPrice } = useMarkPrice(symbol);
   const ticker = useTickerStream(symbol);
+  const { quote_dp, base_dp } = useContext(SymbolContext);
 
-  console.log(ticker);
+  // console.log(ticker);
 
   return (
-    <div className="h-[280px] p-4">
+    <div className="h-[280px] p-4 relative">
       <table className="w-full text-sm">
         <tr className="h-[28px]">
           <td className="text-base-contrast/50">Mark Price</td>
           <td className="text-right">
-            <Numeral>{ticker?.mark_price}</Numeral>
+            <Numeral precision={quote_dp}>{ticker?.mark_price}</Numeral>
           </td>
         </tr>
         <tr className="h-[28px]">
           <td className="text-base-contrast/50">Index Price</td>
           <td className="text-right">
-            <Numeral>{ticker?.index_price}</Numeral>
+            <Numeral precision={quote_dp}>{ticker?.index_price}</Numeral>
           </td>
         </tr>
         <tr className="h-[28px]">
           <td className="text-base-contrast/50">24h Volume</td>
-          <td className="text-right">{ticker?.["24h_volumn"]}</td>
+          <td className="text-right">
+            <Numeral.total
+              rule="human"
+              precision={base_dp}
+              price={ticker?.["24h_close"]}
+              quantity={ticker?.["24h_volumn"]}
+            />
+          </td>
         </tr>
         <tr className="h-[28px]">
           <td className="text-base-contrast/50">24h High</td>
-          <td className="text-right">{ticker?.["24h_high"]}</td>
+          <td className="text-right">
+            <Numeral precision={quote_dp}>{ticker?.["24h_high"]}</Numeral>
+          </td>
         </tr>
         <tr className="h-[28px]">
           <td className="text-base-contrast/50">24h Low</td>
-          <td className="text-right">{ticker?.["24h_low"]}</td>
+          <td className="text-right">
+            <Numeral precision={quote_dp}>{ticker?.["24h_low"]}</Numeral>
+          </td>
         </tr>
         <tr className="h-[28px]">
           <td className="text-base-contrast/50">Open Interest</td>