+++ packages/component/src/tab/tabContent.tsx
@@ -3,7 +3,12 @@ import { TabContext } from "./tabContext";
 import { cx } from "class-variance-authority";
 import { twMerge } from "tailwind-merge";
 
-export const TabContent: FC<PropsWithChildren> = (props) => {
+export interface TabContentProps {
+  keepAlive?: boolean;
+}
+
+export const TabContent: FC<PropsWithChildren<TabContentProps>> = (props) => {
+  const { keepAlive } = props;
   const { contentVisible } = useContext(TabContext);
 
   return (
@@ -18,7 +23,7 @@ export const TabContent: FC<PropsWithChildren> = (props) => {
         console.log("onTransitionEnd");
       }}
     >
-      <div className="overflow-hidden">{props.children}</div>
+      <div className="overflow-hidden relative">{props.children}</div>
     </div>
   );
 };