+++ packages/hooks/src/orderly/useMarketTradeStream.ts
@@ -4,7 +4,7 @@ import { useWS } from "../useWS";
 import { useEffect, useState } from "react";
 
 export interface MarketTradeStreamOptions {
-  level?: number;
+  limit?: number;
 }
 
 export const useMarketTradeStream = (
@@ -16,35 +16,69 @@ export const useMarketTradeStream = (
   }
 
   const [trades, setTrades] = useState<API.Trade[]>([]);
+  const [isLoading, setIsLoading] = useState<boolean>(false);
 
-  const { level = 20 } = options;
+  const { limit = 50 } = options;
 
-  const { isLoading } = useQuery<API.Trade[]>(
-    `/v1/public/market_trades?symbol=${symbol}&limit=${level}`,
-    {
-      onSuccess: (data) => {
-        // console.log("trades ^^^^^^", data);
-        if (Array.isArray(data)) {
-          setTrades(() => data);
-        }
-        return data;
-      },
-    }
-  );
+  // const { isLoading } = useQuery<API.Trade[]>(
+  //   `/v1/public/market_trades?symbol=${symbol}&limit=${level}`,
+  //   {
+  //     onSuccess: (data) => {
+  //       // console.log("trades ^^^^^^", data);
+  //       if (Array.isArray(data)) {
+  //         setTrades(() => data);
+  //       }
+  //       return data;
+  //     },
+  //   }
+  // );
+
+  // const [requestData, setRequestData] = useState<API.Trade[]>([]);
 
   const ws = useWS();
 
   useEffect(() => {
+    setIsLoading(true);
+    setTrades(() => []);
+    ws.onceSubscribe(
+      {
+        id: `${symbol}@trade`,
+        event: "request",
+        params: {
+          type: "trade",
+          symbol: symbol,
+          limit,
+        },
+      },
+      {
+        onMessage: (data: any) => {
+          setIsLoading(false);
+          setTrades(() => data);
+        },
+      }
+    );
+  }, [symbol]);
+
+  useEffect(() => {
+    if (trades.length <= 0) return;
+
     const unsubscript = ws.subscribe(`@${symbol}/@trade`, {
       onMessage: (data: any) => {
-        console.log("ws: trade", data);
+        // console.log("ws: trade topic", data);
+        setTrades((prev) => {
+          const arr = [data, ...prev];
+          if (arr.length > limit) {
+            arr.pop();
+          }
+          return arr;
+        });
       },
     });
 
     return () => {
       unsubscript?.();
     };
-  }, []);
+  }, [symbol, trades]);
 
   return { data: trades, isLoading };
 };