+++ packages/futures/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @orderly.network/futures
 
+## 0.0.28
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@0.0.27
+  - @orderly.network/utils@0.0.21
+
 ## 0.0.27
 
 ### Patch Changes