+++ packages/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @orderly.network/react
 
+## 0.0.21
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/components@0.0.27
+  - @orderly.network/hooks@0.0.28
+
 ## 0.0.20
 
 ### Patch Changes